LPS-97123 Moving extractor api to info
