+++ modules/apps/asset/asset-auto-tagger-opennlp-impl/src/main/java/com/liferay/asset/auto/tagger/opennlp/internal/OpenNLPTextAssetAutoTagProvider.java
@@ -17,10 +17,10 @@ package com.liferay.asset.auto.tagger.opennlp.internal;
 import com.liferay.asset.auto.tagger.AssetAutoTagProvider;
 import com.liferay.asset.auto.tagger.opennlp.api.OpenNLPDocumentAssetAutoTagger;
 import com.liferay.asset.auto.tagger.opennlp.internal.configuration.OpenNLPDocumentAssetAutoTaggerCompanyConfiguration;
-import com.liferay.asset.auto.tagger.opennlp.internal.extractor.TextExtractor;
-import com.liferay.asset.auto.tagger.opennlp.internal.extractor.TextExtractorProvider;
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRenderer;
+import com.liferay.info.extractor.TextExtractor;
+import com.liferay.info.extractor.TextExtractorProvider;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
