+++ modules/apps/asset/asset-auto-tagger-opennlp-impl/src/main/java/com/liferay/asset/auto/tagger/opennlp/internal/extractor/BlogsEntryTextExtractor.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.asset.auto.tagger.opennlp.internal.extractor;
+
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.info.extractor.TextExtractor;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Alicia García
+ * @author Alejandro Tardín
+ */
+@Component(service = TextExtractor.class)
+public class BlogsEntryTextExtractor implements TextExtractor<BlogsEntry> {
+
+	@Override
+	public String getClassName() {
+		return BlogsEntry.class.getName();
+	}
+
+	@Override
+	public String getText(BlogsEntry blogsEntry) {
+		return blogsEntry.getContent();
+	}
+
+}
