+++ modules/apps/info/info-api/bnd.bnd
@@ -7,6 +7,7 @@ Export-Package:\
 	com.liferay.info.display.contributor.field,\
 	com.liferay.info.display.request.attributes.contributor,\
 	com.liferay.info.display.url.provider,\
+	com.liferay.info.extractor,\
 	com.liferay.info.pagination,\
 	com.liferay.info.provider,\
 	com.liferay.info.renderer,\
