+++ modules/apps/info/info-api/src/main/resources/com/liferay/info/extractor/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
